requirements.txt
pandas==2.2.3
openpyxl==3.1.5
pyyaml-6.0.2
pytest-8.3.5